Conflicts:
	BaseLib/ArrayPrimitiveTemplate.h
	BaseLib/ArraySimple.h
	BaseLib/ArrayTemplate.h
	BaseLib/DebugLog.cpp
	BaseLib/Define.h
	BaseLib/FastMemset.h
	BaseLib/LinkListTemplate.h
	BaseLib/PointerRemapper.h
	BaseLib/Quaternion.h
	CoreLib/CalcOperator.cpp
	CoreLib/ListenSocket.cpp
	CoreLib/TypeParser.cpp
	CppParserLib/AbstractSyntaxTree.cpp
	CppParserLib/Preprocessor.cpp
	CppParserLib/Project.cpp
	InputLib/WinDirectInput.cpp
	LanguageLib/NodeElement.h
	ShapeLib_Old/ConvexHull.h
	ShapeLib_Old/Distances.cpp
	ShapeLib_Old/LineSegment.h
	ShapeLib_Old/Plane.h
	ShapeLib_Old/Shape.h
	ShapeLib_Old/ShapeTransformer.cpp
	ShapeLib_Old/Sphere.h
	ShapeLib_Old/VertexArray.cpp
	ShapeLib_Old/VertexArray.h
	StandardLib/ArrayCommonObject.h
	StandardLib/MapStringUnknown.h
	StandardLib/ObjectSerialiser.h
	SupportLib/ImageWriter.cpp
	SupportLib/WinText.cpp
	WindowLib/ComponentPointers.cpp
	WindowLib/FlipContainer.h
	WindowLib/Label.cpp
	WindowLib/Widget.cpp
	WorldLib/Interpolator.h
	WorldLib/InterpolatorArray.h

+++ BaseLib/ASCIITree.cpp
@@ -177,7 +177,7 @@ int CASCIITree::PrivateAdd(CASCIINode* pcNode, int iNode, int iLen, char* szText
 
 		pcChild = mcNodes.Add();
 		pcChild->Init(pcNode, (int)c);
-		pcNode->mapcChildren[(int)c] = pcChild;
+		pcNode->mapcChildren[c] = pcChild;
 		pcNode = pcChild;
 	}
 
@@ -196,7 +196,7 @@ int CASCIITree::PrivateAdd(CASCIINode* pcNode, int iNode, int iLen, char* szText
 
 	pcChars->mlliID = lliID;
 
-	pcNode->maiWordEnds[(int)c] = iIndex;
+	pcNode->maiWordEnds[c] = iIndex;
 	return iIndex;
 }
 
@@ -496,7 +496,7 @@ void CASCIITree::PrivateMoveLastToRemoved(EASCIITreeRemoveStyle eStyle, CASCIINo
 
 		c = pszidEnd->GetChar(pszidEnd->Length()-1);
 		c -= ASCII_NODE_START_CHAR;
-		pcEndLastNode->maiWordEnds[(int)c] = iWordEnd;
+		pcEndLastNode->maiWordEnds[c] = iWordEnd;
 	}
 	else
 	{
