+++ CppParserLib/Preprocessor_Windows.cpp
@@ -40,7 +40,7 @@ void CPreprocessor::InitPlatformSpecific(void)
 	
 #ifdef __cplusplus
 	sz.Init();
-	sz.Append((int)__cplusplus); // FIXME: This should be a long int on MinGW/GCC
+	sz.Append(__cplusplus);
 	AddDefine("__cplusplus", sz.Text());
 	sz.Kill();
 #endif
